requirements-dev.txt
ruff
pytest
pytest-django